bot/requirements.txt
# Discord bot
interactions.py>=5.13.0

# HTTP client (for API communication)
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
